feat: Create MVP Activity list on WearOS
Adds a dynamic list of activities to the Wear
app's main screen which pulls directly from the
phone.

Also includes a refresh button to trigger a manual
resync without needing to reboot the Wear app.

The UI leaves a few things to be lacking,
particularly support for built-in icons instead of
just emojis and text. Additionally, pressing an
activity doesn't yet do anything.

advances thehale/SimpleTimeTracker-WearOS#3