feat(wear): Add navigation layer
This is the last big commit for creating the initial, guiding
project structure.

Replaces the monolithic `MainActivity` with a variety of
narrowly focused Composables which assemble nicely into
a full application.

In particular, this commit shows off the ability to press an activity
which opens a new screen with chips for that activity's tags.

A major milestone towards #5