refactor: Rename `affectCurrentActivities` to `setCurrentActivities`
"affect" is too generic of a word that fails to
precisely describe the type of change it causes in
the system. "set" is more precise and better
matches the naming of the underlying WearRPCClient
method it eventually calls.

Also removed the unnecessary currying of the
callback.
Now it's (Array<Current>Activity>) -> Unit
instead of
(Array<Current>Activity>) -> (Array<Current>Activity>) -> Unit