feat: Support tag selection mode setting
On the phone, the user can request for activities to be started
immediately after picking one tag, or after picking multiple.

With this commit, that behavior is reflected in the WearOS app.

Additionally, the new Submit button on the TagList allows users
to start the activity from the TagsScreen, but without picking
any tags, so it further satisfies thehale/SimpleTimeTracker-WearOS#13

closes thehale/SimpleTimeTracker-WearOS#16
closes thehale/SimpleTimeTracker-WearOS#17